changed tensor2im() image type

+++ util/util.py
@@ -41,7 +41,7 @@ def find_class_in_module(target_cls_name, module):
     return cls
 
 
-def tensor2im(input_image, imtype=np.uint8):
+def tensor2im(input_image, imtype=np.float64):
     """"Converts a Tensor array into a numpy image array.
 
     Parameters:
